cdc/requirements.txt
selenium==4.11.2
pymongo[srv]==3.12.0
pydantic==2.5.3
pydantic_core==2.14.6
google-cloud-logging==3.10.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
pydantic-settings==2.2.1
certifi==2024.2.2
pika==1.3.1
bytewax==0.19.1
unstructured==0.11.8
unstructured-client==0.22.0